+++ spec/dummy/app/controllers/categories_controller.rb
@@ -8,7 +8,7 @@ class CategoriesController < ApplicationController
   end
 
   def create
-    @category = Category.new(params[:category])
+    @category = Category.new(category_params)
     if @category.save
       redirect_to categories_path
     else
@@ -22,7 +22,7 @@ class CategoriesController < ApplicationController
 
   def update
     @category = Category.find(params[:id])
-    if @category.update_attributes(params[:category])
+    if @category.update_attributes(category_params)
       redirect_to categories_path
     else
       render :edit
@@ -35,5 +35,11 @@ class CategoriesController < ApplicationController
     respond_to do |format|
       format.json { render :json => categories.map(&:to_token) }
     end
+  end
+
+  private
+
+  def category_params
+    params.require(:category).permit(:name, :parent_id)
   end
 end