+++ spec/dummy/app/controllers/products_controller.rb
@@ -8,7 +8,7 @@ class ProductsController < ApplicationController
   end
 
   def create
-    @product = Product.new(params[:product])
+    @product = Product.new(product_params)
     if @product.save
       redirect_to products_path
     else
@@ -22,10 +22,16 @@ class ProductsController < ApplicationController
 
   def update
     @product = Product.find(params[:id])
-    if @product.update_attributes(params[:product])
+    if @product.update(product_params)
       redirect_to products_path
     else
       render :edit
     end
+  end
+
+  private
+
+  def product_params
+    params.require(:product).permit(:name, :category_ids)
   end
 end