+++ spec/dummy/app/models/category.rb
@@ -1,5 +1,4 @@
 class Category < ActiveRecord::Base
-  attr_accessible :name, :parent_id
 
   has_many :categories, :foreign_key => :parent_id
   belongs_to :parent, :foreign_key => :parent_id, :class_name => "Category"