+++ spec/dummy/app/models/item.rb
@@ -1,5 +1,3 @@
 class Item < ActiveRecord::Base
-  attr_accessible :category_id, :name
-
   belongs_to :category
 end